made url file and transferred url class to it
